taxi-project/enricher/app/requirements.txt
requests==2.28.1
python-dotenv==0.20.0
fastapi==0.8.0
uvicorn[standard]==0.18.3
sqlalchemy==1.4.40
google-cloud
google-cloud-bigquery
google-cloud-storage